+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased](https://github.com/orisai/scheduler/compare/2.1.1...v2.x)
 
+### Fixed
+
+- `ProcessJobExecutor`
+	- start and end times passed from job process to main process are not shifted by offset of current timezone from UTC
+
 ## [2.1.1](https://github.com/orisai/scheduler/compare/2.1.0...2.1.1) - 2024-06-21
 
 ### Added